pyproject.toml
[project]
name = "nlp-vectorizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "docling>=2.68.0",
    "pillow>=11.3.0",
    "torch>=2.9.1",
    "transformers>=4.57.5",
]
